DSP MATLAB Project #1
In this MATLAB program we are sampling a analog(continous) signal, making the result a discrete time signal. This is one of the fundamental mathematical procedures in digital signal processing. Here in this project, the MATLAB code is developed to provide the Analog to Digital conversion of the signals and the output the waveforms of the respective discrete signals. Note that this process is known as a simple Analog to Digital converter (A/D or ADC).